bloodBackend/bloodaccounts/urls.py
from django.urls import path
from .views import RegisterView
from rest_framework_simplejwt.views import TokenObtainPairView
from .views import MyTokenObtainPairView

urlpatterns = [
    path("auth/register/", RegisterView.as_view()),
    path("auth/login/", TokenObtainPairView.as_view()),
    path("auth/login/", MyTokenObtainPairView.as_view(), name="token_obtain_pair"),

]
